+++ include/MathUtils.h
@@ -7,6 +7,9 @@
 
 #pragma once
 
+// C++ Standard Library includes
+#include <cmath>
+
 namespace astu {
 
     template <typename T>
@@ -43,7 +46,19 @@ namespace astu {
             return phi * static_cast<T>(180.0 / 3.14159265358979323846);
         }
 
-    };
+        /**
+         * Calculates the diagonal of a two dimensional square.
+         * 
+         * @param width the width of the square
+         * @param height    the height of the square
+         * @return the length of the diagonal of the specified square
+         * @tparam  the numerical type
+         */
+        template <typename T>
+        static T RectDiag2D(T width, T height) {
+            return std::sqrt(width * width + height * height);
+        }
 
+    };
 
 } // end of namespace